[MODIFY] 클래스 네이밍 전체적으로 변경
