--- lib/24_03_29/dto/movie_dto.dart
@@ -1,71 +1,0 @@
-class MovieDto {
-  final bool? adult;
-  final String? backdropPath;
-  final List<num>? genreIds;
-  final num? id;
-  final String? originalLanguage;
-  final String? originalTitle;
-  final String? overview;
-  final num? popularity;
-  final String? posterPath;
-  final String? releaseDate;
-  final String? title;
-  final bool? video;
-  final num? voteAverage;
-  final num? voteCount;
-
-  MovieDto({
-    required this.adult,
-    required this.backdropPath,
-    required this.genreIds,
-    required this.id,
-    required this.originalLanguage,
-    required this.originalTitle,
-    required this.overview,
-    required this.popularity,
-    required this.posterPath,
-    required this.releaseDate,
-    required this.title,
-    required this.video,
-    required this.voteAverage,
-    required this.voteCount,
-  });
-
-  factory MovieDto.fromJson(dynamic json) {
-    return MovieDto(
-        adult: json['adult'],
-        backdropPath: json['backdrop_path'],
-        genreIds:
-            json['genre_ids'] != null ? json['genre_ids'].cast<num>() : [],
-        id: json['id'],
-        originalLanguage: json['original_language'],
-        originalTitle: json['original_title'],
-        overview: json['overview'],
-        popularity: json['popularity'],
-        posterPath: json['poster_path'],
-        releaseDate: json['release_date'],
-        title: json['title'],
-        video: json['video'],
-        voteAverage: json['vote_average'],
-        voteCount: json['vote_count']);
-  }
-
-  Map<String, dynamic> toJson() {
-    final map = <String, dynamic>{};
-    map['adult'] = adult;
-    map['backdrop_path'] = backdropPath;
-    map['genre_ids'] = genreIds;
-    map['id'] = id;
-    map['original_language'] = originalLanguage;
-    map['original_title'] = originalTitle;
-    map['overview'] = overview;
-    map['popularity'] = popularity;
-    map['poster_path'] = posterPath;
-    map['release_date'] = releaseDate;
-    map['title'] = title;
-    map['video'] = video;
-    map['vote_average'] = voteAverage;
-    map['vote_count'] = voteCount;
-    return map;
-  }
-}
