+++ lib/24_03_29/mapper/movie_info_mapper.dart
@@ -1,17 +1,23 @@
 import 'package:learn_dart_together/24_03_29/dto/movie_info_dto.dart';
-import 'package:learn_dart_together/24_03_29/mapper/date_mapper.dart';
-import 'package:learn_dart_together/24_03_29/mapper/movie_mapper.dart';
-import 'package:learn_dart_together/24_03_29/model/date.dart';
 import 'package:learn_dart_together/24_03_29/model/movie_info.dart';
 
 extension MovieInfoDtoToMovieInfo on MovieInfoDto {
   MovieInfo toMovieInfo() {
     return MovieInfo(
-      dates: dates?.toDate() ?? Date(),
-      page: page ?? 0,
-      results: results?.map((e) => e.toMovie()).toList() ?? [],
-      totalPages: totalPages ?? 0,
-      totalResults: totalResults ?? 0,
+      adult: adult ?? false,
+      backdropPath: backdropPath ?? 'unknown',
+      genreIds: genreIds ?? [],
+      id: id ?? 0,
+      originalLanguage: originalLanguage ?? 'unknown',
+      originalTitle: originalTitle ?? 'unknown',
+      overview: overview ?? 'unknown',
+      popularity: popularity ?? 0,
+      posterPath: posterPath ?? 'unknown',
+      releaseDate: releaseDate ?? 'unknown',
+      title: title ?? 'unknown',
+      video: video ?? false,
+      voteAverage: voteAverage ?? 0,
+      voteCount: voteCount ?? 0,
     );
   }
 }
