--- lib/24_03_29/model/movie.dart
@@ -1,79 +1,0 @@
-import 'package:collection/collection.dart';
-
-class Movie {
-  final bool adult;
-  final String backdropPath;
-  final List<num> genreIds;
-  final num id;
-  final String originalLanguage;
-  final String originalTitle;
-  final String overview;
-  final num popularity;
-  final String posterPath;
-  final String releaseDate;
-  final String title;
-  final bool video;
-  final num voteAverage;
-  final num voteCount;
-
-  Movie({
-    required this.adult,
-    required this.backdropPath,
-    required this.genreIds,
-    required this.id,
-    required this.originalLanguage,
-    required this.originalTitle,
-    required this.overview,
-    required this.popularity,
-    required this.posterPath,
-    required this.releaseDate,
-    required this.title,
-    required this.video,
-    required this.voteAverage,
-    required this.voteCount,
-  });
-
-  @override
-  String toString() {
-    return 'Movie{adult: $adult, backdropPath: $backdropPath, genreIds: $genreIds, id: $id, originalLanguage: $originalLanguage, originalTitle: $originalTitle, overview: $overview, popularity: $popularity, posterPath: $posterPath, releaseDate: $releaseDate, title: $title, video: $video, voteAverage: $voteAverage, voteCount: $voteCount}';
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is Movie &&
-          runtimeType == other.runtimeType &&
-          adult == other.adult &&
-          backdropPath == other.backdropPath &&
-          genreIds.equals(other.genreIds) &&
-          id == other.id &&
-          originalLanguage == other.originalLanguage &&
-          originalTitle == other.originalTitle &&
-          overview == other.overview &&
-          popularity == other.popularity &&
-          posterPath == other.posterPath &&
-          releaseDate == other.releaseDate &&
-          title == other.title &&
-          video == other.video &&
-          voteAverage == other.voteAverage &&
-          voteCount == other.voteCount;
-
-  @override
-  int get hashCode =>
-      adult.hashCode ^
-      backdropPath.hashCode ^
-      genreIds
-          .reduce((value, element) => value.hashCode ^ element.hashCode)
-          .toInt() ^
-      id.hashCode ^
-      originalLanguage.hashCode ^
-      originalTitle.hashCode ^
-      overview.hashCode ^
-      popularity.hashCode ^
-      posterPath.hashCode ^
-      releaseDate.hashCode ^
-      title.hashCode ^
-      video.hashCode ^
-      voteAverage.hashCode ^
-      voteCount.hashCode;
-}
